requirements.txt
annotated-types==0.7.0

anyio==4.9.0

asgiref==3.10.0

Authlib==1.6.0

bcrypt==4.3.0

blinker==1.9.0
certifi==2025.6.15

cffi==1.17.1

charset-normalizer==3.4.2

click==8.2.0

cmake==4.1.2
colorama==0.4.6

cryptography==45.0.5

Django==5.2.7

dnspython==2.8.0

et_xmlfile==2.0.0
Faker==38.0.0

fastapi==0.115.12

Flask==3.0.3

Flask-Bcrypt==1.0.1

flask-cors==6.0.0
Flask-Login==0.6.3

Flask-Mail==0.9.1

Flask-MySQLdb==2.0.0

Flask-SQLAlchemy==3.1.1

greenlet==3.2.3
gunicorn==21.2.0

h11==0.16.0

idna==3.10

itsdangerous==2.2.0

Jinja2==3.1.6
joblib==1.5.1

MarkupSafe==3.0.2

mysqlclient==2.2.7

nltk==3.9.1

numpy==2.2.6
oauthlib==3.3.1

opencv-contrib-python==4.12.0.88

openpyxl==3.1.5

packaging==25.0

pandas==2.2.3
passlib==1.7.4

pillow==11.3.0

psycopg2==2.9.11

psycopg2-binary==2.9.11

pycparser==2.22
pydantic==2.11.5

pydantic_core==2.33.2

PyJWT==2.10.1

pymongo==4.15.4

python-dateutil==2.9.0.post0
python-dotenv==1.1.1

pytz==2025.2

pywin32==311

regex==2024.11.6

reportlab==4.4.3
requests==2.32.4

requests-oauthlib==2.0.0

scikit-learn==1.6.1

scipy==1.15.3

setuptools==80.9.0
six==1.17.0

sniffio==1.3.1

SQLAlchemy==2.0.41

sqlparse==0.5.3

starlette==0.46.2
textblob==0.19.0

threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
Werkzeug==3.1.3
wheel==0.45.1
xlrd==2.0.2